Port_AI_WebApp/requirements.txt
streamlit
pandas
numpy
scikit-learn
joblib
matplotlib
seaborn
pypdf
google-generativeai
lightgbm
xgboost